Update GetPost and GetPostByCategoryId

+++ ShopSecondHand/Data/ResponseModels/PostResponse/GetPostResponse.cs
@@ -7,7 +7,7 @@ namespace ShopSecondHand.Data.ResponseModels.PostResponse
         public Guid Id { get; set; }
         public string Title { get; set; }
         public string Description { get; set; }
-        public float? Price { get; set; }
+        public double? Price { get; set; }
         public string ImageUrl { get; set; }
         public Guid? AccountId { get; set; }
         public DateTime? CreateAt { get; set; }
